Add API to create series in Tobira
This adds:
- a function to send a `create` request to Opencast
- the graphQL endpoint to call that function from the frontend

The function needs to be given ACL and a title as parameters,
and can be given a description. It then constructs the
metadata json that is expected by the Opencast endpoint and
appends this and the serialized ACL information.

The Opencast endpoint returns a new Opencast ID for the
series, which in conjunction with the title and ACL  can be
used to "prefigure" the series in Tobira.
Doing so allows a new series entry to be shown in the
"My series" table without having to wait for sync.