Implement workaround for series deletion
So this is super annoying. The endpoint in opencast will
start `republish metadata` workflows for each event in the
series. This takes roughly 10 second **per video**.
Not in parallel but.. in.. series (pun intended? idk).

So long story short, it could take several minutes to return,
at which point it probably just responds with a `503`.
The series is still being deleted though, so I don't know how
to handle that case.

Anyway, this workaround will send the request, mark the series
as deleted and then just return without waiting. If it fails,
it will **eventually** become apparent, but until then it's just
showing as `deletion pending`, or sth. Actually, depending on
the configured interval, it will already show `deletion failed`
after a couple of minutes or sooner, when the "deletion process"
is still running in Opencast. So all in all a poor solution but
idk what else to do here other than an Opencast rewrite™.