Remove `useMetadataForm` hook
All this did was provide a timeout option and some states
that weren't even used in all cases. That really wasn't
worth the added complexity.

The timeout was moved to the `SubmitButtonWithStatus` and
when the states are really needed, they can be added in
the parent component that uses them.