Add player shortcuts
These are mostly the same ones that are
used in Youtube videos and on similar platforms,
working within the constraints of the paella player.
Unfortunately, one does not simply replace paella's
icons (I don't think it's imposible though), but
in the interest of time this will have to do without
any icons (that would be shown when a shortcut is
activated).

Note: This needs the next commit to work properly on multi-video
pages.